requirements.txt
Flask==3.1.0
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
celery==5.4.0
redis==5.2.1
requests==2.32.3
flask-login==0.6.3
flask-swagger-ui==4.11.1
Werkzeug==3.1.3
PyJWT==2.10.1
psycopg2==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
loguru==0.7.3